requirements.txt
Flask
opencv-python
Pillow
pytesseract
